chore: Move api-router to helpers folder
The `api-router` was located in the wrong directory. As it serves
as a helper interface, it has been moved to the `helpers` folder
for better organization and maintainability. This change improves
the project's structure by grouping related helper utilities
together.
